Minor SA bug fixed included from trunk.
git-svn-id: file:///mnt/stora_backup/satt/backup-before-new-install/var/svn/repos/software/EasyLocalpp/tags/2.0-pre@208 1ceef749-5ca8-4b7e-9fa7-ff5341897c31

+++ src/runners/SimulatedAnnealing.hh
@@ -54,7 +54,7 @@ public:
 	void Print(std::ostream& os = std::cout) const;
   void SetStartTemperature(double st)  { start_temperature = st; }
   void SetCoolingRate(double cr)  { cooling_rate = cr; }
-  void SetNeighborsSampled(double ns)  { neighbors_sampled = ns; }
+  void SetNeighborsSampled(unsigned int ns)  { neighbors_sampled = ns; }
 
 
 protected:
